+++ lib/taxbear/cli.rb
@@ -4,7 +4,12 @@ module Taxbear
   class CLI < Thor
     desc "login", "Set API key used to authenticate with TaxJar"
     def login
-      token = ask("What is your TaxJar API token?")
+      valid_token = false
+      while !valid_token
+        token = ask("What is your TaxJar API token?")
+        valid_token = Taxjar.validate_token(token)
+        say_nope if !valid_token
+      end
       Config.save_token(token)
       print_success "Success! You are now ready to access the TaxJar API."
     end
@@ -27,6 +32,22 @@ module Taxbear
     def print_success(message)
       success_message = "✔︎ #{message}"
       puts set_color(success_message, :green)
+    end
+
+    def say_nope
+      print_error(random_no)
+    end
+
+    def random_no
+      [
+        "Nope, that's not correct.",
+        "Sorry, not quite right.",
+        "Maybe you should check the website? https://app.taxjar.com/smartcalcs",
+        "A for effort… but no.",
+        "I'll give you a hint, it's a bunch of random unguessable letters/numbers.",
+        "Good try, but unfortunately wrong.",
+        "Hmm… that doesn't seem to be working. Maybe something different?"
+      ].sample
     end
   end
 end