+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.3 (2016-10-18)
 - Move login to a separate command.
+- Validate token provided against TaxJar API.
 
 ## 0.1.2 (2016-10-18)
 - Improve CLI help.